fix: move `zDAO` type in files
