+++ src/types.ts
@@ -42,6 +42,19 @@ export interface Config {
 export type zNA = string;
 export type zDAOId = string;
 
+export interface zDAO {
+  id: zDAOId; // Global zDAO identifier
+  zNA: zNA; // Linked zNA
+  title?: string; // zDAO title
+  creator: string; // Creator wallet address
+  owners: string[]; // Owner wallet addresses
+  avatar?: string; // Avatar uri (https link)
+  network: string; // Chain id
+  // strategies?: any; // only used for snapshot, @todo
+  safeAddress: string; // Gnosis Safe address
+  votingToken: string; // Voting token address
+}
+
 export interface CreateZDAODto {
   // zNA
   zNA: zNA;
